Allow the plugin to read a default file
Reading a default file for options allows for CM to deploy a custom
password for the system to scout.

This changed required using a newer supported library, upgrading
from mysql to mysql2 gem.

Unit tests aren't passing, it doesn't appear the mysql2 gem supports
stubbing, and trying to call the count method without having a real
database handle causes it to segfault. This only impacts test not
real usage.

This was reported upstream:
https://github.com/brianmario/mysql2/issues/610